Add *ALL* 3x3 puzzles that can be solved in 20 or fewer moves
Generated with a quick and dirty script. Created a depth-first search
that starts with the goalState and continues trying moves that result in
states that have not already been seen.

It seems that any 3x3 puzzle can be solved in 31 moves or fewer.

This is a huge amount of data, but I figure it might be easier for us to
start with *all* puzzles and maybe prune the number later.